+++ OCKSample/AppDelegate.swift
@@ -40,7 +40,6 @@ class AppDelegate: UIResponder, UIApplicationDelegate, ObservableObject {
 
     let isSyncingWithCloud = true // True to sync with ParseServer, False to Sync with iOS Watch
     var isFirstAppOpen = true
-    var isFirstLogin = false
     private var sessionDelegate: SessionDelegate!
     private lazy var watch = OCKWatchConnectivityPeer()
     private(set) var parseRemote: ParseRemote!
@@ -60,7 +59,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate, ObservableObject {
     func application(_ application: UIApplication,
                      didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
 
-        // Parse-server setup
+        // Parse-Server setup
         PCKUtility.setupServer(fileName: Constants.parseConfigFileName) { _, completionHandler in
             completionHandler(.performDefaultHandling, nil)
         }
@@ -84,9 +83,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate, ObservableObject {
                     NotificationCenter.default.post(.init(name: Notification.Name(rawValue: Constants.requestSync)))
                 }
             }
-
         } else {
-
             // When syncing directly with watchOS, we do not care about login and need to setup remotes
             setupRemotes()
             Task {
@@ -109,12 +106,10 @@ class AppDelegate: UIResponder, UIApplicationDelegate, ObservableObject {
                 }
             }
         }
-
         return UISceneConfiguration(name: "Default Configuration", sessionRole: connectingSceneSession.role)
     }
 
-    func application(_ application: UIApplication, didDiscardSceneSessions sceneSessions: Set<UISceneSession>) {
-    }
+    func application(_ application: UIApplication, didDiscardSceneSessions sceneSessions: Set<UISceneSession>) {}
 
     func application(_ application: UIApplication,
                      didRegisterForRemoteNotificationsWithDeviceToken deviceToken: Data) {
@@ -122,7 +117,10 @@ class AppDelegate: UIResponder, UIApplicationDelegate, ObservableObject {
             await Utility.updateInstallationWithDeviceToken(deviceToken)
         }
     }
+}
 
+// MARK: Helpers
+extension AppDelegate {
     func resetAppToInitialState() {
         NotificationCenter.default.post(.init(name: Notification.Name(rawValue: Constants.storeDeinitialized)))
         do {
@@ -136,7 +134,6 @@ class AppDelegate: UIResponder, UIApplicationDelegate, ObservableObject {
             Logger.appDelegate.error("Error deleting OCKStore: \(error.localizedDescription)")
         }
         isFirstAppOpen = true
-        isFirstLogin = true
         storeManager = .init(wrapping: OCKStore(name: Constants.noCareStoreName, type: .inMemory))
         healthKitStore = nil
         parseRemote = nil
@@ -145,7 +142,6 @@ class AppDelegate: UIResponder, UIApplicationDelegate, ObservableObject {
 
     func setupRemotes(uuid: UUID? = nil) {
         do {
-
             if isSyncingWithCloud {
                 guard let uuid = uuid else {
                     Logger.appDelegate.error("Error in setupRemotes, uuid is nil")
@@ -168,10 +164,12 @@ class AppDelegate: UIResponder, UIApplicationDelegate, ObservableObject {
                 sessionDelegate = LocalSessionDelegate(remote: watch, store: store)
             }
 
+            // Setup communication with watch
             WCSession.default.delegate = sessionDelegate
             WCSession.default.activate()
 
             guard let currentStore = store else {
+                Logger.appDelegate.error("Should have OCKStore")
                 return
             }
             healthKitStore = OCKHealthKitPassthroughStore(store: currentStore)
@@ -181,55 +179,6 @@ class AppDelegate: UIResponder, UIApplicationDelegate, ObservableObject {
             storeManager = OCKSynchronizedStoreManager(wrapping: coordinator)
         } catch {
             Logger.appDelegate.error("Error setting up remote: \(error.localizedDescription)")
-        }
-    }
-}
-
-extension AppDelegate: ParseRemoteDelegate {
-
-    func didRequestSynchronization(_ remote: OCKRemoteSynchronizable) {
-        NotificationCenter.default.post(.init(name: Notification.Name(rawValue: Constants.requestSync)))
-    }
-
-    func successfullyPushedDataToCloud() {
-        if self.isFirstAppOpen {
-            self.isFirstLogin = false
-            self.isFirstAppOpen = false
-            NotificationCenter.default.post(.init(name: Notification.Name(rawValue: Constants.reloadView)))
-        }
-    }
-
-    func remote(_ remote: OCKRemoteSynchronizable, didUpdateProgress progress: Double) {
-        let progressPercentage = Int(progress * 100.0)
-        NotificationCenter.default.post(.init(name: Notification.Name(rawValue: Constants.progressUpdate),
-                                              userInfo: [Constants.progressUpdate: progressPercentage]))
-    }
-
-    func chooseConflictResolution(conflicts: [OCKEntity], completion: @escaping OCKResultClosure<OCKEntity>) {
-
-        // https://github.com/carekit-apple/CareKit/issues/567
-        // Workaround to handle deleted and re-added outcomes.
-        // Always prefer updates over deletes.
-        let outcomes = conflicts.compactMap { conflict -> OCKOutcome? in
-            if case let .outcome(outcome) = conflict {
-                return outcome
-            } else {
-                return nil
-            }
-        }
-
-        if outcomes.count == 2,
-           outcomes.contains(where: { $0.deletedDate != nil}),
-           let added = outcomes.first(where: { $0.deletedDate == nil}) {
-
-            completion(.success(.outcome(added)))
-            return
-        }
-
-        if let first = conflicts.first {
-            completion(.success(first))
-        } else {
-            completion(.failure(.remoteSynchronizationFailed(reason: "Error, none selected for conflict")))
         }
     }
 }
